move all navigation to its own components in a separate folder (#106)
