+++ navigation/RootStackParamList.tsx
@@ -1,0 +1,9 @@
+import FutureUserEvent from '../types/futureUserEvent';
+
+export type RootStackParamList = {
+  Start: undefined;
+  Map: undefined;
+  Event: undefined;
+  EventForm: { event: FutureUserEvent | null; };
+  Profile: undefined;
+};
